Зарефакторил ProfileViewController
